Sơn Hà [ADD UPDATE] update lock user when user login wrong password over 5 times,  add class test for unit testing

+++ src/main/java/com/group6/swp391/service/UserService.java
@@ -48,7 +48,7 @@ public interface UserService {
 
     public int getPartDate(Date date, int calendarPart);
 
-    public int calculateSecondIn5Minute(User user);
+    public long calculateSecondIn5Minute(User user);
 
     public boolean checkEmailOrPhone(String s);
